Wayland support
Detect that we're using Wayland (tricky because we don't have type
bindings)
Then get the wayland display and surface and use that to create
the swapsurface.
And ask for the Vulkan wayland extension.
Hack: Recreating every 10 cycles, need to make it actually recreate
on resize
Rendering always at centre of display